refactored most of the parsing using the examples on https://json-ld.org/playground/ as a tests to check the effectiveness of the approach.
I got several ideas on how to apply this. Howeover, I first would like to a full functionality.

For now, unmarshalling is not in the scope, also because I do not know what exactly it would mean.
Even if it sounds easy to parse back a POJO out of the json-ld, I'm not sure how to handle annotations backwards.

However, several points are still missing. I'll open issues afterwards:

- expanded, flattered and framed mods.
- @List, @graph and several other interesting keywords.
- cleanup